Make quick sort stable (#7)
Quicksort is a comparison sort, meaning it can sort items of any type for
which a total order is defined. Various efficient implementations of quick
sort are not a stable sort, meaning the relative order of equal sort items
is not preserved. This change would instead implement a stable sort. i.e.
the relative position of item with the same value would remain same after
sorting.

Reported and patched by Chien-Chi Liao <maeglin89273@gmail.com>